+++ src/pages/home/contents/DashboardPage/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useSelectedTagMemosManager } from './useSelectedTagMemosManager';
 export { default as useTagsManager } from './useTagManager';
+export { default as useDragAndDropManager } from './useDragDropManager';
